Migrate /user/{id}
Make PostList component server-side
