+++ frontend/src/api/posts/edge.ts
@@ -23,3 +23,9 @@ export const getCategoriesLocal = (): Promise<string[]> =>
     if (!res.ok) throw new Error("Network response was not ok")
     return res.json()
   })
+
+export const getUserPostsLocal = (user_id: number): Promise<Post[]> =>
+  fetch(`${process.env.FORUM_BACKEND_PRIVATE_URL}/api/user/${user_id}/posts`).then((res) => {
+    if (!res.ok) throw new Error("Network response was not ok")
+    return res.json()
+  })
