+++ frontend/src/api/users/edge.ts
@@ -1,0 +1,7 @@
+import { User } from "@/custom"
+
+export const getUserLocal = (id: number): Promise<User> =>
+  fetch(`${process.env.FORUM_BACKEND_PRIVATE_URL}/api/user/${id}`).then((res) => {
+    if (!res.ok) throw new Error("Network response was not ok")
+    return res.json()
+  })
