+++ frontend/src/components/posts/list.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import Link from "next/link"
 import { FC } from "react"
